requirements.txt
gitdb==4.0.11
GitPython==3.1.40
javalang==0.13.0
six==1.16.0
smmap==5.0.1
